src/requirements.txt
django==1.11.2
pymongo==3.4.0
redis==2.10.5
adslproxy==2.0.0
scrapy==1.3.3
scrapyd==1.2.0
beautifulsoup4
selenium
scrapy_redis
lxml==3.8.0
readability-lxml
